# openapi.yaml
openapi: 3.0.3
info:
  title: Subscription Management API
  version: 1.0.0
  description: |
    Frontend APIs use JWT. SDK APIs use API Key from /sdk/auth/login.

servers:
  - url: http://localhost:8080
    description: Local

paths:
  /api/customer/login:
    post:
      summary: Customer login (frontend)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Returns JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { type: object, properties: { token: { type: string }, user: { type: object } } }
        '401':
          description: Invalid credentials

  /api/admin/login:
    post:
      summary: Admin login (frontend)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Returns JWT token
        '401':
          description: Invalid credentials

  /api/customer/signup:
    post:
      summary: Customer sign up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string }
                name: { type: string }
      responses:
        '201':
          description: Created, returns JWT
        '409':
          description: Email already registered

  /api/v1/customer/profile:
    get:
      summary: Get customer profile (JWT)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Customer profile

  /api/v1/customer/subscription:
    get:
      summary: Get current subscription (JWT)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current or null
    post:
      summary: Request subscription (body.sku)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [sku]
              properties:
                sku: { type: string }
      responses:
        '201':
          description: Request created

  /api/v1/admin/customers:
    get:
      summary: List customers (admin JWT)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paginated list

  /api/v1/admin/subscription-packs:
    get:
      summary: List packs (admin JWT)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List of packs
    post:
      summary: Create pack (admin JWT)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, sku, price, validity_months]
              properties:
                name: { type: string }
                description: { type: string }
                sku: { type: string }
                price: { type: number }
                validity_months: { type: integer, minimum: 1, maximum: 12 }
      responses:
        '201':
          description: Created

  /api/v1/admin/subscriptions:
    get:
      summary: List subscriptions (admin JWT)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema: { type: string, enum: [requested, approved, active, inactive, expired] }
      responses:
        '200':
          description: Paginated list

  /api/v1/admin/subscriptions/{id}/approve:
    post:
      summary: Approve subscription (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Approved

  /api/v1/admin/subscriptions/{id}/assign:
    post:
      summary: Assign subscription (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Assigned

  /api/v1/admin/subscriptions/{id}/deactivate:
    post:
      summary: Deactivate subscription (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Deactivated

  /sdk/auth/login:
    post:
      summary: SDK login (returns API key)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Returns api_key and user

  /sdk/v1/subscription:
    get:
      summary: Get current subscription (X-API-Key)
      security: [{ apiKey: [] }]
      responses:
        '200':
          description: Current subscription or null

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key